project/movie_recommend/back/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.9
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
